PR feedback: Extract auto-resolve into a shared core module, rename, simplify return type
