+++ src/commands/quality-gate/status/scope.ts
@@ -21,11 +21,9 @@
 // Resolves which branch or pull request `quality-gate status` reports on
 
 import { CommandFailedError, InvalidOptionError } from '@/core/commands/command-error.ts';
-import { resolveCurrentGitBranch } from '@/core/host/git/branch.ts';
-import logger from '@/core/observability/logger.ts';
+import { autoResolvePullRequest } from '@/core/pull-request-auto-resolve.ts';
 import { BranchesClient } from '@/core/server/branches.ts';
 import type { SonarHttpClient } from '@/core/server/http-client.ts';
-import { PullRequestsClient } from '@/core/server/pull-requests.ts';
 
 export interface QualityGateStatusScopeOptions {
   branch?: string;
@@ -71,7 +69,7 @@ export async function resolveQualityGateScope(
 
   const branchesClient = new BranchesClient(client);
   const [autoDetected, branches] = await Promise.all([
-    resolveAutoPullRequest(client, projectKey),
+    autoResolvePullRequest(client, projectKey),
     branchesClient.listBranches(projectKey),
   ]);
   if (autoDetected) {
@@ -92,35 +90,4 @@ export async function resolveQualityGateScope(
     });
   }
   return { queryParams: {}, scope: { kind: 'default', value: defaultBranch.name } };
-}
-
-// Never throws — undefined (no git branch, lookup failed, no/ambiguous match) means the caller falls back to the default branch.
-async function resolveAutoPullRequest(
-  client: SonarHttpClient,
-  projectKey: string,
-): Promise<{ pullRequest: string; branch: string } | undefined> {
-  const branch = await resolveCurrentGitBranch(process.cwd());
-  if (!branch) {
-    return undefined;
-  }
-
-  const pullRequests = await tryListPullRequests(client, projectKey);
-  const matches = pullRequests?.filter((pr) => pr.branch === branch) ?? [];
-  if (matches.length !== 1) {
-    return undefined;
-  }
-
-  return { pullRequest: matches[0].key, branch };
-}
-
-async function tryListPullRequests(
-  client: SonarHttpClient,
-  projectKey: string,
-): Promise<Awaited<ReturnType<PullRequestsClient['listPullRequests']>>> {
-  try {
-    return await new PullRequestsClient(client).listPullRequests(projectKey);
-  } catch (err) {
-    logger.debug(`Pull request auto-detection skipped for '${projectKey}'`, err);
-    return null;
-  }
 }
